+++ lib/overcommit/hook/post-rewrite/base.rb
@@ -1,0 +1,10 @@
+require 'forwardable'
+
+module Overcommit::Hook::PostRewrite
+  # Functionality common to all post-rewrite hooks.
+  class Base < Overcommit::Hook::Base
+    extend Forwardable
+
+    def_delegators :@context, :amend?, :rebase?
+  end
+end